requirements.txt
websockets
requests
pyspark
kafka-python-ng
flask
